Adding changes to use logical replication pg connector with a flag

+++ yb-voyager/cmd/exportData.go
@@ -362,14 +362,17 @@ func exportData() bool {
 
 		if changeStreamingIsEnabled(exportType) {
 			log.Infof("live migration complete, proceeding to cutover")
-			if isTargetDBExporter(exporterRole) {
+			msr, err := metaDB.GetMigrationStatusRecord()
+			if err != nil {
+				utils.ErrExit("get migration status record: %v", err)
+			}
+			if isTargetDBExporter(exporterRole) && !msr.UseLogicalReplicationYBConnector {
 				err = ybCDCClient.DeleteStreamID()
 				if err != nil {
 					utils.ErrExit("failed to delete stream id after data export: %v", err)
 				}
 			}
 			if exporterRole == SOURCE_DB_EXPORTER_ROLE {
-				msr, err := metaDB.GetMigrationStatusRecord()
 				if err != nil {
 					utils.ErrExit("get migration status record: %v", err)
 				}
